fix: allow WS_HOST to be undefined
fix: invalidate localcontext instead of parent context

fix: create an endpoint to get the context from the mocks

fix: get mock context

fix: save context

fix: update position of toolbox

fix: use getLocalcontext data

fix: get token from cache

fix: local context

fix: disabled test for postmessage

chore: update lockfile

fix: post setting

fix: permission

fix: move mocking solutions in their own folder and cleanup

fix: apply PR comments
