+++ src/api/index.ts
@@ -2,3 +2,4 @@ export * from './apps';
 export * from './appData';
 export * from './appAction';
 export * from './appSetting';
+export * from './mocks';
