+++ src/mockServer/types.ts
@@ -1,0 +1,3 @@
+import { Request } from 'miragejs';
+
+export type ExternalUrls = (((req: Request) => unknown) | string)[];
